# docs/openapi.json
{"openapi":"3.1.0","servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/scans":{"get":{"tags":["scan-controller"],"operationId":"getScanHistory","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageScanSummaryResponse"}}}}}},"post":{"tags":["scan-controller"],"operationId":"submitScan","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"image":{"type":"string","format":"binary"}},"required":["image"]}}}},"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScanSubmitResponse"}}}}}}},"/api/v1/auth/resend-verification":{"post":{"tags":["auth-controller"],"operationId":"resendVerification","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendVerificationRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}}}}},"/api/v1/auth/forgot-password":{"post":{"tags":["auth-controller"],"operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/users/profile":{"get":{"tags":["user-controller"],"operationId":"getCurrentUserProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentUserProfileResponse"}}}}}},"patch":{"tags":["user-controller"],"operationId":"updateUserProfile","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentUserProfileResponse"}}}}}}},"/test":{"get":{"tags":["ai-test-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FoodSafetyResponse"}}}}}}},"/api/v1/users/me":{"get":{"tags":["user-controller"],"operationId":"getCurrentUserSummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentUserSummaryResponse"}}}}}}},"/api/v1/scans/{scanId}":{"get":{"tags":["scan-controller"],"operationId":"getScanResult","parameters":[{"name":"scanId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScanResultResponse"}}}}}}},"/api/v1/diseases":{"get":{"tags":["disease-controller"],"operationId":"getAllDiseases","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DiseaseResponse"}}}}}}}},"/api/v1/diseases/{id}":{"get":{"tags":["disease-controller"],"operationId":"getDiseaseById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiseaseResponse"}}}}}}},"/api/v1/allergies":{"get":{"tags":["allergy-controller"],"operationId":"getAllAllergies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AllergyResponse"}}}}}}}},"/api/v1/allergies/{id}":{"get":{"tags":["allergy-controller"],"operationId":"getAllergyById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AllergyResponse"}}}}}}}},"components":{"schemas":{"ScanSubmitResponse":{"type":"object","properties":{"scanId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["PROCESSING","COMPLETED","FAILED"]}}},"ResendVerificationRequest":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"RegisterRequest":{"type":"object","properties":{"firstName":{"type":"string","maxLength":100,"minLength":0},"lastName":{"type":"string","maxLength":100,"minLength":0},"email":{"type":"string","maxLength":255,"minLength":0},"username":{"type":"string","maxLength":100,"minLength":0},"password":{"type":"string","maxLength":128,"minLength":8},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"heightCm":{"type":"number","minimum":0.0},"weightKg":{"type":"number","minimum":0.0},"allergies":{"type":"array","items":{"type":"integer","format":"int32"}},"diseases":{"type":"array","items":{"type":"integer","format":"int32"}}},"required":["email","firstName","lastName","password","username"]},"RegisterResponse":{"type":"object","properties":{"message":{"type":"string"},"requiresEmailVerification":{"type":"boolean"}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"UpdateProfileRequest":{"type":"object","properties":{"firstName":{"type":"string","maxLength":100,"minLength":0},"lastName":{"type":"string","maxLength":100,"minLength":0},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"heightCm":{"type":"number","minimum":0.0},"weightKg":{"type":"number","minimum":0.0},"allergies":{"type":"array","items":{"type":"integer","format":"int32"}},"diseases":{"type":"array","items":{"type":"integer","format":"int32"}}}},"AllergyResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"CurrentUserProfileResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"heightCm":{"type":"number"},"weightKg":{"type":"number"},"allergies":{"type":"array","items":{"$ref":"#/components/schemas/AllergyResponse"}},"diseases":{"type":"array","items":{"$ref":"#/components/schemas/DiseaseResponse"}},"updatedAt":{"type":"string","format":"date-time"}}},"DiseaseResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"FlaggedIngredient":{"type":"object","properties":{"ingredient":{"type":"string"},"reason":{"type":"string"},"type":{"type":"string","enum":["ALLERGY","CHRONIC_CONDITION"]},"name":{"type":"array","items":{"type":"string"}}}},"FoodSafetyResponse":{"type":"object","properties":{"verdict":{"type":"string","enum":["SAFE","UNSAFE","CAUTION"]},"flaggedIngredients":{"type":"array","items":{"$ref":"#/components/schemas/FlaggedIngredient"}},"summary":{"type":"string"}}},"CurrentUserSummaryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageScanSummaryResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ScanSummaryResponse"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"unpaged":{"type":"boolean"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"ScanSummaryResponse":{"type":"object","properties":{"scanId":{"type":"string","format":"uuid"},"imageUrl":{"type":"string"},"verdict":{"type":"string","enum":["SAFE","UNSAFE","CAUTION"]},"scannedAt":{"type":"string","format":"date-time"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}},"ScanResultResponse":{"type":"object","properties":{"scanId":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["PROCESSING","COMPLETED","FAILED"]},"scannedAt":{"type":"string","format":"date-time"},"imageUrl":{"type":"string"},"foodSafetyResponse":{"$ref":"#/components/schemas/FoodSafetyResponse"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Enter your JWT token here. Do not include 'Bearer ' in front of it.","scheme":"bearer","bearerFormat":"JWT"}}}}